[step3] refactor: 패키지 이동
