+++ src/main/java/racing/inputter/Inputter.java
@@ -1,0 +1,6 @@
+package racing.inputter;
+
+public interface Inputter {
+
+    int inputIntValue();
+}
