+++ src/main/java/racing/random/RandomGenerator.java
@@ -1,0 +1,6 @@
+package racing.random;
+
+public interface RandomGenerator {
+
+    int getRandomIntValue();
+}
