+++ src/main/java/racing/random/RealRandomGenerator.java
@@ -1,0 +1,21 @@
+package racing.random;
+
+import java.util.Random;
+
+public class RealRandomGenerator implements RandomGenerator {
+
+    private final Random random;
+    private final int bound;
+
+    public RealRandomGenerator(int bound) {
+
+        this.random = new Random();
+        this.bound = bound;
+    }
+
+    @Override
+    public int getRandomIntValue() {
+
+        return random.nextInt(bound);
+    }
+}
