--- src/main/java/step2/racing/inputter/Inputter.java
@@ -1,6 +1,0 @@
-package step2.racing.inputter;
-
-public interface Inputter {
-
-    int inputIntValue();
-}
