--- src/main/java/step2/racing/random/RandomGenerator.java
@@ -1,6 +1,0 @@
-package step2.racing.random;
-
-public interface RandomGenerator {
-
-    int getRandomIntValue();
-}
