+++ src/test/java/racing/random/StubRandomGenerator.java
@@ -1,0 +1,17 @@
+package racing.random;
+
+public class StubRandomGenerator implements RandomGenerator {
+
+    private int fixedNumber;
+
+    public StubRandomGenerator(int fixedNumber) {
+
+        this.fixedNumber = fixedNumber;
+    }
+
+    @Override
+    public int getRandomIntValue() {
+
+        return fixedNumber;
+    }
+}
