--- src/test/java/step2/racing/random/StubRandomGenerator.java
@@ -1,17 +1,0 @@
-package step2.racing.random;
-
-public class StubRandomGenerator implements RandomGenerator {
-
-    private int fixedNumber;
-
-    public StubRandomGenerator(int fixedNumber) {
-
-        this.fixedNumber = fixedNumber;
-    }
-
-    @Override
-    public int getRandomIntValue() {
-
-        return fixedNumber;
-    }
-}
